Bootstrap async main runtime
