+++ Cargo.toml
@@ -61,6 +61,7 @@ serde_json = { version = "1.0.149", features = ["preserve_order"] }
 serde_test = { version = "1.0.177" }
 static_assertions = { version = "1.1.0" }
 toml = { version = "1.1.2" }
+tokio = { version = "1.52.3", features = ["macros", "rt", "time"] }
 unic-ucd-category = { version = "0.9" }
 unicode-ident = { version = "1.0.24" }
 unicode_names2 = { version = "2.0.0" }
